+++ GrabMaterials/GrabMaterialsMod.cs
@@ -19,7 +19,7 @@ using UnityEngine;
 namespace GrabMaterialsMod
 {
 
-	[BepInPlugin(GrabMaterialsMod.ModGuid, "Grab Materials", "2.4.0")]
+	[BepInPlugin(GrabMaterialsMod.ModGuid, "Grab Materials", "2.4.1")]
 	[BepInProcess("valheim.exe")]
 	public class GrabMaterialsMod : BaseUnityPlugin
 	{